gulpfile polish and separation of copy and inject tasks to a file for main client generator
move copy tasks to file, move inject tasks to file, add injects and copy files to index.js, remove unused imports, polish
