+++ generators/client/templates/gulp/_config.js
@@ -3,6 +3,7 @@
 module.exports = {
     app: '<%= MAIN_SRC_DIR %>',
     dist: '<%= DIST_DIR %>',
+    swaggerDist: '<%= DIST_DIR %>swagger-ui/',
     test: '<%= TEST_SRC_DIR %>'<% if(useSass) { %>,
     scss: '<%= MAIN_SRC_DIR %>scss/',
     sassSrc: '<%= MAIN_SRC_DIR %>scss/**/*.{scss,sass}',
